Modify fonts, modify audio header files

+++ native/cocos/platform/openharmony/modules/CanvasRenderingContext2DDelegate.h
@@ -48,7 +48,7 @@ public:
     using Point   = ccstd::array<float, 2>;
     using Vec2    = ccstd::array<float, 2>;
     using Size    = ccstd::array<float, 2>;
-    using Color4F = ccstd::array<uint8_t, 4>;
+    using Color4F = ccstd::array<float, 4>;
     using TextAlign = ICanvasRenderingContext2D::TextAlign;
     using TextBaseline = ICanvasRenderingContext2D::TextBaseline;
